Cut out the middle man, use the simpler tree.
Implementing IsForeign and AsForeign on SimpleTree allows this to be
used directly, rather than translated through the MyTree record type.

The record type was originally there because it matched the structure of
the data in the d3 tree example, but this structure is now only
contained in the typeclass instances.